finishes integrating context, in processs of writing tests for context
