+++ client/src/__tests__/overview.test.js
@@ -1,0 +1,2 @@
+import App from '../App';
+import ProductOverview from '../components/overview/ProductOverview';
