+++ client/src/components/R&R/Reviews.jsx
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { useState, useContext } from 'react';
+import { ProductContext } from '../../context';
 
 const Reviews = () => {
+  const {reviews, reviewsMeta, reviewSort, setReviewSort } = useContext(ProductContext)
+
   return <div>This is the Ratings and Reviews widget!</div>;
 };
